# internal/api/docs/swagger.yaml
definitions:
  api.AddBookmarkInput:
    properties:
      label:
        type: string
      link:
        type: string
    required:
    - label
    - link
    type: object
  api.GetAnimalOutput:
    properties:
      animal:
        $ref: '#/definitions/db.Animal'
    type: object
  api.GetAnimalsOutput:
    properties:
      animals:
        items:
          $ref: '#/definitions/db.Animal'
        type: array
    type: object
  api.GetBookmarksOutput:
    properties:
      bookmarks:
        items:
          $ref: '#/definitions/db.Bookmark'
        type: array
    type: object
...
  api.UpdateRateOfGainInput:
    properties:
      beginning_date:
        type: string
      beginning_weight:
        type: number
      end_date:
        type: string
      end_weight:
        type: number
    required:
    - beginning_date
    - beginning_weight
    - end_date
    - end_weight
    type: object
...
  api.UpsertAnimalInput:
    properties:
      animal_cost:
        type: string
      birth_date:
        type: string
      dam_breed:
        type: string
      name:
        type: string
      projectid:
        type: string
      purchase_date:
        type: string
      quality_grade:
        type: string
      sale_price:
        type: string
      sire_breed:
        type: string
      species:
        type: string
      yield_grade:
        type: string
    required:
    - animal_cost
    - birth_date
    - dam_breed
    - name
    - projectid
    - purchase_date
    - quality_grade
    - sale_price
    - sire_breed
    - species
    - yield_grade
    type: object
...
  db.Animal:
    properties:
      animal_cost:
        type: string
      beginning_date:
        type: string
      beginning_weight:
        type: number
      birth_date:
        type: string
      created:
        type: string
      dam_breed:
        type: string
      end_date:
        type: string
      end_weight:
        type: number
      id:
        type: string
      name:
        type: string
      projectid:
        type: string
      purchase_date:
        type: string
      quality_grade:
        type: string
      sale_price:
        type: string
      sire_breed:
        type: string
      species:
        type: string
      updated:
        type: string
      userid:
        type: string
      yield_grade:
        type: string
    type: object
  db.Bookmark:
    properties:
      created:
        type: string
      id:
        type: string
      label:
        type: string
      link:
        type: string
      updated:
        type: string
      userid:
        type: string
    type: object
...
info:
  contact: {}
  title: 4H Record Books API
paths:
  /animal:
    get:
      consumes:
      - application/json
      description: Gets all of a user's animals for a given project
      parameters:
      - description: Project ID
        in: query
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetAnimalsOutput'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get animals by project
      tags:
      - Animal
    post:
      consumes:
      - application/json
      description: Adds an animal to a user's personal records
      parameters:
      - description: Animal information
        in: body
        name: UpsertAnimalInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertAnimalInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Add an animal
      tags:
      - Animal
  /animal/{animalId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's animal given the animal ID
      parameters:
      - description: Animal ID
        in: path
        name: animalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes an animal
      tags:
      - Animal
    get:
      consumes:
      - application/json
      description: Get a user's animal by ID
      parameters:
      - description: Animal ID
        in: path
        name: animalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetAnimalOutput'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get an animal
      tags:
      - Animal
    put:
      consumes:
      - application/json
      description: Updates a user's animal information
      parameters:
      - description: Animal ID
        in: path
        name: animalId
        required: true
        type: string
      - description: Animal information
        in: body
        name: UpsertAnimalInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertAnimalInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update an animal
      tags:
      - Animal
  /bookmarks:
    get:
      consumes:
      - application/json
      description: Returns an array of all the user's bookmarks, queried using JWT
        claims
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetBookmarksOutput'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get all of a user's bookmarks
      tags:
      - User Bookmarks
    post:
      consumes:
      - application/json
      description: |-
        Adds a bookmark to a user's personal records.
        The new bookmark can not have the same link as another of the user's bookmarks
      parameters:
      - description: Bookmark information
        in: body
        name: AddBookmarkInput
        required: true
        schema:
          $ref: '#/definitions/api.AddBookmarkInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "409":
          description: Conflict
      security:
      - ApiKeyAuth: []
      summary: Adds a bookmark
      tags:
      - User Bookmarks
  /bookmarks/{bookmarkId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's bookmark given the bookmark ID
      parameters:
      - description: Bookmark ID
        in: path
        name: bookmarkId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes a bookmark
      tags:
      - User Bookmarks
...
  /rate-of-gain/{animalId}:
    put:
      consumes:
      - application/json
      description: Updates a user's animal rate of gain information
      parameters:
      - description: Animal ID
        in: path
        name: animalId
        required: true
        type: string
      - description: Animal rate of gain information
        in: body
        name: UpdateRateOfGainInput
        required: true
        schema:
          $ref: '#/definitions/api.UpdateRateOfGainInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update an animal's rate of gain
      tags:
      - Animal
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"